Refactoring toolbar menupopover
