+++ react/components/EXPERIMENTAL_Table/BulkActions.tsx
@@ -13,6 +13,7 @@ import {
   BULK_ACTIONS_TRANSITION,
 } from './constants'
 import useTableBulkActions from './stateContainers/bulkActions'
+import { MenuAction } from './Toolbar/PopoverMenu'
 
 const BulkActions: FC<BulkActionsProps> = ({
   texts,
