--- react/components/EXPERIMENTAL_Table/Toolbar/Menu/context.ts
@@ -1,17 +1,0 @@
-import { createContext, useContext } from 'react'
-
-interface MenuContext {
-  isBoxVisible: boolean
-  setBoxVisible: (isBoxVisible: boolean) => void
-}
-
-const MenuContext = createContext<MenuContext>(null)
-const { Provider: MenuProvider } = MenuContext
-
-const useMenuContext = () => {
-  const context = useContext(MenuContext)
-  return context
-}
-
-export default MenuContext
-export { MenuProvider, useMenuContext }
