+++ react/components/EXPERIMENTAL_Table/typings.d.ts
@@ -1,13 +1,3 @@
-type MenuAction = {
-  label: string
-  onClick: Function
-  toggle?: {
-    checked: boolean
-    semantic: boolean
-  }
-  id?: number | string
-}
-
 type Density = 'low' | 'medium' | 'high'
 type Size = 'small' | 'regular' | 'large'
 type Alignment = 'left' | 'right'
